requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Sloan\Documents\GitHub\Spam686-Python\spam686.py: 89
# C:\Users\Sloan\Documents\GitHub\Spam686-Python\test.py: 7,10
Pillow == 8.2.0

# C:\Users\Sloan\Documents\GitHub\Spam686-Python\spam686.py: 9
numba == 0.53.1

# C:\Users\Sloan\Documents\GitHub\Spam686-Python\spam686.py: 3,88
# C:\Users\Sloan\Documents\GitHub\Spam686-Python\test.py: 8
numpy == 1.20.2

# C:\Users\Sloan\Documents\GitHub\Spam686-Python\test.py: 9
oct2py == 5.2.0
